requirements.txt
pyyaml
mysql-connector-python
requests
